+++ foundrytools_cli_2/cli/commands/fix/__init__.py
@@ -1,0 +1,3 @@
+from foundrytools_cli_2.cli.commands.fix.cli import cli
+
+__all__ = ["cli"]
