--- foundrytools_cli_2/cli/fix/tasks/nbsp_missing.py
@@ -1,9 +1,0 @@
-from foundrytools import Font
-
-
-def main(font: Font) -> bool:
-    """
-    Fixes the missing non-breaking space glyph by double mapping the space glyph.
-    """
-    font.t_cmap.add_missing_nbsp()
-    return font.t_cmap.is_modified
