--- foundrytools_cli_2/cli/os_2/__init__.py
@@ -1,3 +1,0 @@
-from foundrytools_cli_2.cli.os_2.cli import cli
-
-__all__ = ["cli"]
